Split dense output tests into separate files
